requirements.txt
pyheif
Pillow
click
